All_extract_1_2_3/1 - scrap_2 paruvendu -- Source 1/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
Automat==24.8.1
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.26.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.30.0
isodate==0.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.1
matplotlib-inline==0.1.7
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.1
Protego==0.4.0
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.10.1
pyodbc==5.2.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tldextract==5.1.3
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
w3lib==2.3.1
wcwidth==0.2.13
wget==3.2
xxhash==3.5.0
yarl==1.18.3
zope.interface==7.2